+++ CHANGES.md
@@ -3,6 +3,7 @@
 ## Version 0.5
 
 * Switched from Freebase to The Movie Database
+* Added buffer limits for Android streaming requirements
 
 ## Version 0.4